consumer/requirements.txt
aiokafka==0.11.0
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.1
pyee==12.0.0
python-ffmpeg==2.0.12
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.2
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.5
transformers==4.44.2
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets 
wsproto 
youtube-dl 
yt-dlp 